Fix for https://github.com/finagle/finch/issues/895: rename  `io.finch.syntax.scala` to `io.finch.syntax.scalaFutures` (per @vkostyukov)
