--- core/src/main/scala/io/finch/syntax/scala/package.scala
@@ -1,3 +1,0 @@
-package io.finch.syntax
-
-package object scala extends EndpointMappers with ScalaToTwitterFuture
