+++ core/src/test/scala/io/finch/syntax/ScalaMapperSyntaxSpec.scala
@@ -1,9 +1,8 @@
-package io.finch.syntax.scala
+package io.finch.syntax
 
+import io.finch.syntax.scalafutures._
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
-
-import io.finch.syntax.MapperSyntaxSpec
 
 class ScalaMapperSyntaxSpec extends MapperSyntaxSpec {
